requirements.txt
discord.py==2.1.0
emoji_country_flag==1.3.0
flag==0.1.1
Flask==2.0.2
Pillow==9.3.0
pycountry==22.1.10
requests==2.28.1
tabulate==0.8.9
